Add Cobra to support additional commands
